fix(trace): use float32 for confidence warn threshold to match confidence type
The threshold was stored as float64 and cast to float32 at comparison time,
which silently shifted the effective value (e.g. float32(0.85) = 0.850000024,
not 0.85). Cast once at config load instead, keeping the entire comparison
in float32 space where confidence already lives.

Also fix two pre-existing test bugs:
- MCP tools_test passed evidence as []map[string]any instead of a JSON string,
  so GetString returned "" and the handler saw zero evidence
- server_test never set HighConfidenceWarnThreshold, defaulting to 0 and
  triggering warnings on every trace regardless of confidence

+++ internal/model/api.go
@@ -279,8 +279,8 @@ type TraceEvidence struct {
 
 // HighConfidenceWarnings returns warnings when confidence exceeds threshold
 // with zero evidence items. Returns nil when no warning applies.
-func HighConfidenceWarnings(confidence float32, evidenceCount int, threshold float64) []string {
-	if confidence > float32(threshold) && evidenceCount == 0 {
+func HighConfidenceWarnings(confidence float32, evidenceCount int, threshold float32) []string {
+	if confidence > threshold && evidenceCount == 0 {
 		return []string{
 			fmt.Sprintf(
 				"high confidence (%.2g) with no supporting evidence — consider adding evidence items or lowering confidence to 0.4–0.8",
